fix: various optimizations
