+++ package.json
@@ -78,7 +78,6 @@
   "dependencies": {
     "@intlify/shared": "^9.7.0",
     "@intlify/unplugin-vue-i18n": "^1.4.0",
-    "@mizchi/sucrase": "^4.1.0",
     "@nuxt/kit": "^3.7.4",
     "@vue/compiler-sfc": "^3.3.4",
     "debug": "^4.3.4",
@@ -87,11 +86,10 @@
     "is-https": "^4.0.0",
     "knitwork": "^1.0.0",
     "magic-string": "^0.30.4",
-    "mlly": "^1.4.2",
     "pathe": "^1.1.1",
+    "sucrase": "^3.34.0",
     "ufo": "^1.3.1",
     "unplugin": "^1.5.0",
-    "unstorage": "^1.9.0",
     "vue-i18n": "9.6.5",
     "vue-i18n-routing": "^1.1.4"
   },
